+++ jason/src/api/mod.rs
@@ -2,7 +2,6 @@
 
 mod connection;
 mod room;
-pub mod room_state;
 
 use std::{cell::RefCell, rc::Rc};
